+++ components/MessageBoard/MessageItem.tsx
@@ -8,13 +8,16 @@ import { isMobileOnly } from "react-device-detect"
 import { StyleSheet, Text, TouchableOpacity, View } from "react-native"
 import ProfileImage from "../../components/ProfileImage/ProfileImage"
 import { Message, Reply } from "./MessageList"
+import { DM } from "./MessageListDirect"
 import MessageUtils from "./MessageUtils"
 
 type Props = {
-  item: Message | Reply
+  item: Message | Reply | DM
   index: number
   isReply: boolean
   now: Moment
+  isDM?: boolean
+  isMine?: boolean
 }
 
 const style = StyleSheet.create({
@@ -64,7 +67,7 @@ const style = StyleSheet.create({
   },
 })
 const MessageItem = (props: Props) => {
-  const { item, index, isReply, now } = props
+  const { item, index, isReply, now, isDM } = props
   const navigation = useNavigation<StackNavigationProp<any, any>>()
   const convertCommentFromJSONToHTML = (text: string | null) => {
     const errorMarkdown = "<div>" + "Message" + " Can't Be Displayed</div>"
@@ -81,7 +84,6 @@ const MessageItem = (props: Props) => {
   const showProfile = (id: string | undefined) => {
     if (id) navigation?.push("ProfileScreen", { id: id, create: false })
   }
-
   const currentTime = now
   const datePosted = moment(parseInt(item?.when))
   const daysSince = currentTime.diff(datePosted, "days")
@@ -98,8 +100,98 @@ const MessageItem = (props: Props) => {
       ? daysSince + " days ago"
       : null
   }`
+  if (isDM) {
+    return (
+      <>
+        <View
+          style={
+            props.isMine
+              ? {
+                  flexDirection: "row-reverse",
+                  alignSelf: "flex-end",
+                  maxWidth: "65ch",
+                  marginBottom: 16,
+                }
+              : { flexDirection: "row", display: "flex", maxWidth: "65ch", marginBottom: 16 }
+          }
+        >
+          {!props.isMine ? (
+            <View>
+              <ProfileImage size={"small7"} user={item?.userId} />
+            </View>
+          ) : null}
+          <View
+            style={
+              props.isMine
+                ? {
+                    flexDirection: "column",
+                    flex: 1,
+                    justifyContent: "flex-end",
+                    alignItems: "flex-end",
+                  }
+                : {
+                    flexDirection: "column",
+                    flex: 1,
+                  }
+            }
+          >
+            <View
+              style={
+                props.isMine
+                  ? {
+                      backgroundColor: "#FFECEB",
+                      paddingVertical: 8,
+                      paddingHorizontal: 16,
+                      borderRadius: 8,
+                      alignSelf: "flex-end",
+                      flexShrink: 1,
+                    }
+                  : {
+                      backgroundColor: "#EDEBEB",
+                      paddingVertical: 8,
+                      paddingHorizontal: 16,
+                      borderRadius: 8,
+                      alignSelf: "flex-start",
+                      flexShrink: 1,
+                    }
+              }
+            >
+              <Text
+                style={{
+                  color: "#1A0706",
+                  fontFamily: "Graphik-Regular-App",
+                  fontSize: 15,
+                }}
+              >
+                <div
+                  style={{ marginBlockEnd: 0 }}
+                  dangerouslySetInnerHTML={{
+                    __html: convertCommentFromJSONToHTML(item?.content ?? null)
+                      ?.replaceAll("<p>", "")
+                      .replaceAll("</p>", ""),
+                    // need to filter empty elements here
+                  }}
+                ></div>
+              </Text>
+            </View>
+            <Text
+              style={{
+                fontFamily: "Graphik-Regular-App",
+                fontSize: 14,
+                lineHeight: 21,
+                color: "#6A5E5D",
+                marginTop: 16,
+              }}
+            >
+              {moment(parseInt(item?.when)).format("lll")}
+            </Text>
+          </View>
+        </View>
+      </>
+    )
+  }
   return (
-    <View key={index} style={[style.MessageContainer, isReply ? style.ReplyContainer : {}]}>
+    <View style={[style.MessageContainer, isReply ? style.ReplyContainer : {}]}>
       <View>
         <View style={{ flexDirection: "row" }}>
           <View style={{ flexDirection: "column" }}>
