+++ screens/ProfilesScreen/ProfileCard.tsx
@@ -136,7 +136,12 @@ export default function ProfileCard(props: Props) {
               customLabel: { fontFamily: "Graphik-Medium-App" },
             }}
             label={"Message"}
-            action={() => null}
+            action={() => {
+              navigation.push("ConversationScreen", {
+                initialUserID: item?.id,
+                initialUserName: item?.given_name + " " + item?.family_name,
+              })
+            }}
             icon={"Airplane"}
           />
         </View>
